6/25 update
Connection Status button now updates text, but needs ng-class to update
icon. Counters now work, but for some reason do not update if the data
is the same. Need to pass SID between controllers (factory/service).
Tolerances now added, location data is being passed, and fields are
cleared on purge. Input fields now take decimals. Picture placeholder
tab added.

To do:
>for loop recursive upload (+error handlers)
>shrink interface
>factory for SID
>camera plugin
>counter fix (low priority)
>orientation plugin
>push location data to DOM
>EXTENSIVE TESTING

+++ www/js/controllers.js
@@ -22,9 +22,22 @@ angular.module('starter.controllers', [])
             states[Connection.CELL_4G]  = 'Cell 4G connection';
             states[Connection.CELL]     = 'Cell generic connection';
             states[Connection.NONE]     = 'No network connection';
+      console.log('Connection type: ' + states[networkState]);
+      if (states[networkState] = 'Wifi connection') {
+        var connect = true;
+        document.getElementById("connect-text").innerHTML = "Connection Status: WiFi Connection";
+      }
+      else if (states[networkState] = 'No network connection') {
+        document.getElementById("connect-icon").class = "icon ion-close connect-icon";
+        document.getElementById("connect-text").innerHTML = "Connection Status: No connection available";
+      } 
+      else {
+        document.getElementById("connect-icon").class = "icon ion-checkmark connect-icon";
+        document.getElementById("connect-text").innerHTML = "Connection Status: Cellular Data Connection";
+      }
 
-            window.alert('Connection type: ' + states[networkState]); 
-        };
+  };
+
 
 
 })
@@ -41,34 +54,25 @@ angular.module('starter.controllers', [])
   $scope.loc = {};
   $scope.fileName = "";
   $scope.data = "";
+  $scope.counter1 = 0;
+  $scope.counter2 = 0;
+  $scope.counter3 = 0;
+  $scope.counter4 = 0;
 
-//Updates the Session ID and Retrieves Location Data
-  $scope.getSID = function () {
-    var SID = $scope.userSID.data;
-    console.log('Session ID set to: ' + SID);
-    window.alert ('Your new session ID is ' + SID);
-    var pos0ptions = {timeout: 10000, enableHighAccuracy: true};
-    $cordovaFile.createDir(cordova.file.documentsDirectory, SID, false)
-    .then(function (success) {
-      console.log("New directory " + SID + " created.");
-    }, function (error) {
-      window.alert("Error: This Session ID has already been used.");
-      console.log(error);
-    });
-    $cordovaGeolocation.getCurrentPosition(pos0ptions)
+  var pos0ptions = {timeout: 10000, enableHighAccuracy: true};
+  $cordovaGeolocation.getCurrentPosition(pos0ptions)
       .then(function(position){
-        console.log('Latitude: '    + position.coords.latitude           + '\n' +
+        /*console.log('Latitude: '    + position.coords.latitude           + '\n' +
             'Longitude: '         + position.coords.longitude         + '\n' +
             'Altitude: '          + position.coords.altitude          + '\n' +
             'Accuracy: '          + position.coords.accuracy          + '\n' +
             'Altitude Accuracy: ' + position.coords.altitudeAccuracy  + '\n' +
             'Heading: '           + position.coords.heading           + '\n' +
             'Speed: '             + position.coords.speed             + '\n' +
-            'Timestamp: '         + position.timestamp                + '\n');
-
+            'Timestamp: '         + position.timestamp                + '\n');*/
       $scope.loc.lat = position.coords.latitude
       $scope.loc.lon = position.coords.longitude
-      $scope.loc.alt = position.coords.altitude 
+      $scope.loc.alti = position.coords.altitude 
       $scope.loc.acc = position.coords.accuracy
       $scope.loc.head = position.coords.heading
       $scope.loc.DTG = position.coords.timestamp
@@ -78,18 +82,17 @@ angular.module('starter.controllers', [])
       alert('code: '    + error.code    + '\n' +
             'message: ' + error.message + '\n');
     })
-  };
-
 
 //Parent export function: Should save file, tick up counter, and clear field
   $scope.doSave1 = function() {
+    if (($scope.temp.data > 0 && $scope.temp.data < 40) && ($scope.temp.depth > 0 && $scope.temp.depth < 100)) {
     $ionicPlatform.ready(function() {
       var SID = $scope.userSID.data;
       var DTG = Date.now()
-      var data = $scope.temp.data + ', ' + $scope.temp.depth;
+      var data = $scope.temp.data + ', ' + $scope.temp.depth + ', ' + $scope.loc.lat + ', ' + $scope.loc.lon + ', ' + $scope.loc.acc + ', ' + $scope.loc.alti + ', ' + $scope.loc.heading + ', ' + $scope.loc.speed;
       var variable = "temperature";
-      var counter = document.getElementById("counter-1").innerHTML;
-      counter++
+      $scope.counter1 += 1;
+      document.getElementById("counter-1").innerHTML = $scope.counter1;
       $scope.fileName = SID + "_" + variable + "_" + DTG + ".oskit";
       $cordovaFile.writeFile(cordova.file.documentsDirectory, $scope.fileName, data, true)
       .then(function (success) {
@@ -111,16 +114,20 @@ angular.module('starter.controllers', [])
         console.log("ERROR: " + error);
       });
     })
+    } else {
+      window.alert("Your values are not within the tolerances. Please check them and resubmit.")
+    }
   };
 
   $scope.doSave2 = function() {
+    if (($scope.salin.data > 0 && $scope.salin.data < 40) && ($scope.salin.depth > 0 && $scope.salin.depth < 100)) {
     $ionicPlatform.ready(function() {
       var SID = $scope.userSID.data;
       var DTG = Date.now()
-      var data = $scope.salin.data + ', ' + $scope.salin.depth;
+      var data = $scope.salin.data + ', ' + $scope.salin.depth + ', ' + $scope.loc.lat + ', ' + $scope.loc.lon + ', ' + $scope.loc.acc + ', ' + $scope.loc.alti + ', ' + $scope.loc.heading + ', ' + $scope.loc.speed;
       var variable = "salinity";
-      var counter = document.getElementById("counter-2").innerHTML;
-      counter++
+      $scope.counter2 += 1;
+      document.getElementById("counter-2").innerHTML = $scope.counter1;
       $scope.fileName = SID + "_" + variable + "_" + DTG + ".oskit";
       $cordovaFile.writeFile(cordova.file.documentsDirectory, $scope.fileName, data, true)
       .then(function (success) {
@@ -142,16 +149,20 @@ angular.module('starter.controllers', [])
         console.log("ERROR: " + error);
       });
     })
+    } else {
+      window.alert("Your values are not within the tolerances. Please check them and resubmit.")
+    }
   };
 
   $scope.doSave3 = function() {
+    if ($scope.sechi.depth > 0 && $scope.sechi.depth < 100) {
     $ionicPlatform.ready(function() {
       var SID = $scope.userSID.data;
       var DTG = Date.now()
-      var data = $scope.sechi.depth
+      var data = $scope.sechi.depth + ', ' + $scope.loc.lat + ', ' + $scope.loc.lon + ', ' + $scope.loc.acc + ', ' + $scope.loc.alti + ', ' + $scope.loc.heading + ', ' + $scope.loc.speed;
       var variable = "sechi";
-      var counter = document.getElementById("counter-3").innerHTML;
-      counter++
+      $scope.counter1 += 3;
+      document.getElementById("counter-3").innerHTML = $scope.counter1;
       $scope.fileName = SID + "_" + variable + "_" + DTG + ".oskit";
       $cordovaFile.writeFile(cordova.file.documentsDirectory, $scope.fileName, data, true)
       .then(function (success) {
@@ -173,16 +184,19 @@ angular.module('starter.controllers', [])
         console.log("ERROR: " + error);
       });
     })
+    } else {
+      window.alert("Your values are not within the tolerances. Please check them and resubmit.")
+    }
   };
 
   $scope.doSave4 = function() {
     $ionicPlatform.ready(function() {
       var SID = $scope.userSID.data;
       var DTG = Date.now();
-      var data = $scope.weather.data;
+      var data = $scope.weather.data + ', ' + $scope.loc.lat + ', ' + $scope.loc.lon + ', ' + $scope.loc.acc + ', ' + $scope.loc.alti + ', ' + $scope.loc.heading + ', ' + $scope.loc.speed;
       var variable = "weather";
-      var counter = document.getElementById("counter-4").innerHTML;
-      counter++
+      $scope.counter4 += 1;
+      document.getElementById("counter-4").innerHTML = $scope.counter1;
       $scope.fileName = SID + "_" + variable + "_" + DTG + ".oskit";
       $cordovaFile.writeFile(cordova.file.documentsDirectory, $scope.fileName, data, true)
       .then(function (success) {
@@ -205,7 +219,6 @@ angular.module('starter.controllers', [])
       });
     })
   };
-
 
 // Upload function: Should upload all files in directory
   $scope.doUpload = function() {
@@ -234,36 +247,64 @@ angular.module('starter.controllers', [])
       })
     })  
   }
+
 $scope.dofUpload = function () {
   setTimeout(function(){ alert("Files finished uploading."); }, 3000);
 };
+
 })
 
 
 
 
 .controller('AccountCtrl', function($scope, $cordovaDialogs, $cordovaFile) {
-$scope.doRemove = function() {
-  $cordovaDialogs.confirm('Are you sure you want to delete all local files? This action is permanant.', 'Purge Local Cache', ['Delete','Cancel'])
-    .then(function(buttonIndex) {
-      // no button = 0, 'OK' = 1, 'Cancel' = 2
-      var btnIndex = buttonIndex;
-    });
+//Removes SID Directory
+  $scope.doRemove = function() {
+    $cordovaDialogs.confirm('Are you sure you want to delete all local files? This action is permanant.', 'Purge Local Cache', ['Delete','Cancel'])
+      .then(function(buttonIndex) {
+        // no button = 0, 'OK' = 1, 'Cancel' = 2
+        var btnIndex = buttonIndex;
+        if (buttonIndex == 1) {
+          $ionicPlatform.ready(function() {
+            $cordovaFile.removeDir(cordova.file.documentsDirectory, SID)
+            .then(function (success) {
+              window.alert("Files successfully deleted.");
+            }, function (error) {
+              window.alert("Error deleting files.");
+              console.log(error);
+            });
+            var form1 = document.getElementById("temp");
+            form.reset();
+            var form2 = document.getElementById("salin");
+            form.reset(); 
+            var form3 = document.getElementById("sechi");
+            form.reset(); 
+            var form4 = document.getElementById("weather");
+            form.reset();   
+          });
+        }
+        else {
+          return null;
+        }
+      });
+  } 
 
-  if (buttonIndex == 1) {
-    $ionicPlatform.ready(function() {
-      $cordovaFile.removeDir(cordova.file.documentsDirectory, SID)
-      .then(function (success) {
-        window.alert("Files successfully deleted.");
-      }, function (error) {
-        window.alert("Error deleting files.");
-        console.log(error);
-      });  
-    });
-  }
-  else {
-    return null;
-  }
+//Updates the Session ID
+  $scope.getSID = function () {
+    var SID = $scope.userSID.data;
+    console.log('Session ID set to: ' + SID);
+    window.alert ('Your new session ID is ' + SID);
+    var pos0ptions = {timeout: 10000, enableHighAccuracy: true};
+    $cordovaFile.createDir(cordova.file.documentsDirectory, SID, false)
+    .then(function (success) {
+      console.log("New directory " + SID + " created.");
+    }, function (error) {
+      window.alert("Error: This Session ID has already been used.");
+      console.log(error);
+    }); 
+  }; 
 
-}  
 })
+
+.controller('CameraCtrl', function($scope) {
+  })
